Revert "[1.0.5] 랜딩페이지 구현"
