+++ src/components/features/Post/plugins/ClipboardPlugin.ts
@@ -459,6 +459,7 @@ const registerPasteCommand = (editor: LexicalEditor) => {
               const codeNode = $createCodeNode();
               codeNode.setLanguage(language);
               codeNode.append($createTextNode(plainText));
+
               $insertNodes([codeNode]);
             });
