+++ src/pages/Home/Home.tsx
@@ -3,38 +3,33 @@ import { Flex } from '@/components/commons/Flex';
 import { SText } from '@/components/commons/SText';
 import { Spacer } from '@/components/commons/Spacer';
 import { QnAList } from '@/components/features/Home/QnAList';
+import { ReviewSlider } from '@/components/features/Home/ReviewSlider';
 import { CommonLayout } from '@/components/layout/CommonLayout';
 
-import GroupIcon from '@/assets/Landing/group_icon.svg';
-import ManageIcon from '@/assets/Landing/manage_icon.svg';
-import RecordIcon from '@/assets/Landing/record_icon.svg';
-
 import { Fragment } from 'react';
+import { useNavigate } from 'react-router-dom';
 
 import logo from '@/assets/Header/logo.svg';
 import bgCurve from '@/assets/Home/bg-curve.svg';
 import bgRing from '@/assets/Home/bg-ring.svg';
 import bgZigzag from '@/assets/Home/bg-zigzag.svg';
 import faq from '@/assets/Home/faq.svg';
+import instagram from '@/assets/Home/instagram.svg';
 import logoLight from '@/assets/Home/logo-light.svg';
-import { AnimatedLanding } from '@/components/features/Landing/AnimatedLanding';
-import { Banner } from '@/components/features/Landing/Banner';
-import ServiceStrength from '@/components/features/Landing/ServiceStrength';
-import UsageExample from '@/components/features/Landing/UsageExample';
+import name from '@/assets/Home/name-last.svg';
+import notion from '@/assets/Home/notion.svg';
+import { PATH } from '@/routes/path';
 import styled from '@emotion/styled';
 
 export const Home = () => {
-  // const navigate = useNavigate();
-  // const onClickStart = () => navigate(`${PATH.TEAM_RECRUIT}/list`);
+  const navigate = useNavigate();
+  const onClickStart = () => navigate(`${PATH.TEAM_RECRUIT}/list`);
 
   return (
     <Fragment>
       <CommonLayout>
         {/* 민경 */}
         <Section backgroundColor="#fff">
-          <AnimatedLanding />
-        </Section>
-        <Section backgroundColor="#FAFAFF">
           <Decoration src={bgZigzag} top="180px" left="20%" />
           <Decoration src={bgRing} top="150px" right="20%" />
           <Decoration src={bgCurve} top="270px" right="27%" width="60px" />
@@ -58,7 +53,7 @@ export const Home = () => {
             color="#767676"
             textAlign="center"
             fontFamily="Pretendard"
-            fontSize="20px"
+            fontSize="24px"
             fontWeight={300}
             lineHeight="34px"
           >
@@ -69,41 +64,19 @@ export const Home = () => {
             풀이를 공유하며 지속적인 성장을 이끌어내는 커뮤니티입니다.
           </SText>
           <Spacer h={50} />
-          <div style={{ display: 'flex', gap: '27px' }}>
-          <Banner
-            title="팀 스터디 관리"
-            description1="팀원과 함께 목표를 설정하고, 매일"
-            description2="문제 풀이 현황을 공유해보세요."
-            src={ManageIcon}
-          />
-          <Banner
-            title="풀이 기록 & 회고"
-            description1="나만의 풀이를 기록하고, 팀원과 피드백을"
-            description2="주고받으며 회고 문화를 만들어갑니다."
-            src={RecordIcon}
-          />
-          <Banner
-            title="스터디 그룹"
-            description1="목표와 실력이 비슷한 동료들을 찾고,"
-            description2="직접 팀을 구성하거나 기존 팀에 참여할 수 있습니다."
-            src={GroupIcon}
-          />
-          </div>
           {/* TODO: 카드 3개 영역 */}
         </Section>
 
-        <Section backgroundColor="#fff">
+        <Section>
           <SubHeader text="이렇게 사용해보세요!" theme="dark" />
-          <UsageExample />
         </Section>
 
         <Section backgroundColor="#000">
           <SubHeader text="코드몬스터 서비스 소개!" theme="light" />
-          <ServiceStrength />
         </Section>
 
         {/* 지수 */}
-        {/* <Section>
+        <Section>
           <SubHeader text="코드몬스터 서비스 소개!" theme="dark" />
           <Spacer h={40} />
           <SText
@@ -129,7 +102,7 @@ export const Home = () => {
           <ReviewSlider />
           <Spacer h={80} />
           <Ellipse />
-        </Section> */}
+        </Section>
 
         <Section>
           <SText
@@ -157,7 +130,7 @@ export const Home = () => {
           <QnAList />
         </Section>
 
-        {/* <Section backgroundColor="#fff">
+        <Section backgroundColor="#fff">
           <Decoration src={bgZigzag} top="140px" left="20%" />
           <Decoration src={bgRing} top="140px" right="20%" />
           <Decoration src={bgCurve} top="220px" right="27%" width="60px" />
@@ -183,7 +156,7 @@ export const Home = () => {
             <img src={notion} alt="notion" />
           </Flex>
           <Spacer h={240} />
-        </Section> */}
+        </Section>
       </CommonLayout>
     </Fragment>
   );
@@ -208,11 +181,11 @@ const SubHeader = ({
   return (
     <Flex direction="column">
       <Logo src={theme === 'dark' ? logo : logoLight} alt={'logo'} />
-      <Spacer h={12} />
+      <Spacer h={10} />
       <SText
         color={theme === 'dark' ? '#333' : '#D5D5D5'}
         fontFamily="Pretendard"
-        fontSize="28px"
+        fontSize="24px"
         fontWeight={900}
         textAlign="center"
       >
@@ -226,14 +199,14 @@ const Logo = styled.img`
   height: 16px;
 `;
 
-// const Ellipse = styled.div`
-//   width: 900px;
-//   height: 19px;
-//   border-radius: 900px;
-//   background: rgba(212, 212, 212, 0.4);
-//   filter: blur(8px);
-//   border-radius: 50%;
-// `;
+const Ellipse = styled.div`
+  width: 900px;
+  height: 19px;
+  border-radius: 900px;
+  background: rgba(212, 212, 212, 0.4);
+  filter: blur(8px);
+  border-radius: 50%;
+`;
 
 const Decoration = styled.img<{
   top: string;
@@ -250,16 +223,16 @@ const Decoration = styled.img<{
   z-index: 1;
 `;
 
-// const StartButton = styled.button`
-//   display: flex;
-//   align-items: center;
-//   justify-content: center;
-//   height: 74px;
-//   padding: 24px 40px 24px 52px;
-//   gap: 4px;
-//   border-radius: 40px;
-//   background: #333;
-//   color: #fff;
-//   font-size: 36px;
-//   font-weight: 400;
-// `;
+const StartButton = styled.button`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 74px;
+  padding: 24px 40px 24px 52px;
+  gap: 4px;
+  border-radius: 40px;
+  background: #333;
+  color: #fff;
+  font-size: 36px;
+  font-weight: 400;
+`;
